cpu: refactor SIMD mappings and vectorized op functions into separate files
